#36-users-films-retrieve-endpoint all tests passed
